+++ FinancialAdvisor/app/src/main/java/team_18/financialadvisor/PseudoDatabaseEntry.java
@@ -5,7 +5,7 @@ package team_18.financialadvisor;
  */
 public class PseudoDatabaseEntry {
     private final String transactionType;
-    private final float transactionAmount;
+    private final double transactionAmount;
     private final String transactionComment;
     private final int transactionID;
 
@@ -13,7 +13,7 @@ public class PseudoDatabaseEntry {
     PseudoDatabaseEntry previousEntry;
     PseudoDatabaseEntry nextEntry;
 
-    public PseudoDatabaseEntry(String type, float amount, String comment, int ID) {
+    public PseudoDatabaseEntry(String type, double amount, String comment, int ID) {
         transactionType = type;
         transactionAmount = amount;
         transactionComment = comment;
